+++ apps/codeimage/src/core/modules/localFontAccessApi/font.ts
@@ -56,3 +56,15 @@ export function determineFontFaceName(familyStyle: string) {
 
   return face;
 }
+
+export const fontWeightLabelMap: Record<number, string> = {
+  100: 'Thin',
+  200: 'Extra Light',
+  300: 'Light',
+  400: 'Regular',
+  500: 'Medium',
+  600: 'Semi Bold',
+  700: 'Bold',
+  800: 'Extra Bold',
+  900: 'Black',
+};
